Move duplicated styles out of log in and sign up components so it is used once
