+++ src/logged_out/Login.js
@@ -11,7 +11,6 @@ import Typography from '@material-ui/core/Typography';
 import Container from '@material-ui/core/Container';
 import {Form, Formik} from "formik";
 import FormikTextField from "../common/components/FormikTextField";
-import {makeStyles} from "@material-ui/core/styles";
 import Copyright from "../common/components/Copyright";
 import {ROOT_URL, SIGN_UP_URL} from "../common/costants/urls";
 import {Link, withRouter} from "react-router-dom";
@@ -20,26 +19,8 @@ import Snackbar from "@material-ui/core/Snackbar";
 import * as yup from 'yup'
 import {authenticationService} from "../common/services/authenticationService";
 import {withApollo} from "@apollo/react-hoc";
-
-const useStyles = makeStyles((theme) => ({
-    paper: {
-        marginTop: theme.spacing(8),
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-    },
-    avatar: {
-        margin: theme.spacing(1),
-        backgroundColor: theme.palette.secondary.main,
-    },
-    form: {
-        width: '100%', // Fix IE 11 issue.
-        marginTop: theme.spacing(1),
-    },
-    submit: {
-        margin: theme.spacing(3, 0, 2),
-    },
-}));
+import { Paper } from '@material-ui/core';
+import useStyles from './styles'
 
 function Login(props) {
     const {client, history} = props;
@@ -59,9 +40,9 @@ function Login(props) {
     };
 
     return (
-        <Container component="main" maxWidth="xs">
+        <Container maxWidth="xs">
             <CssBaseline/>
-            <div className={classes.paper}>
+            <Paper elevation={3} className={classes.paper}>
                 <Avatar className={classes.avatar}>
                     <LockOutlinedIcon/>
                 </Avatar>
@@ -152,7 +133,7 @@ function Login(props) {
                         </Form>
                     )}
                 </Formik>
-            </div>
+        </Paper>
             <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
                 <MuiAlert onClose={handleClose} severity={alertSeverity} elevation={6} variant="filled">
                     {alertMessage}
