+++ src/logged_out/SignUp.js
@@ -4,9 +4,9 @@ import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
+import { Paper } from '@material-ui/core';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
-import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import Copyright from "../common/components/Copyright";
 import {Form, Formik} from "formik";
@@ -17,26 +17,7 @@ import Snackbar from "@material-ui/core/Snackbar";
 import MuiAlert from '@material-ui/lab/Alert';
 import {LOG_IN_URL, ROOT_URL} from "../common/costants/urls";
 import {Link, withRouter} from "react-router-dom";
-
-const useStyles = makeStyles((theme) => ({
-    paper: {
-        marginTop: theme.spacing(8),
-        display: 'flex',
-        flexDirection: 'column',
-        alignItems: 'center',
-    },
-    avatar: {
-        margin: theme.spacing(1),
-        backgroundColor: theme.palette.secondary.main,
-    },
-    form: {
-        width: '100%', // Fix IE 11 issue.
-        marginTop: theme.spacing(3),
-    },
-    submit: {
-        margin: theme.spacing(3, 0, 2),
-    },
-}));
+import useStyles from './styles'
 
 function SignUp(props) {
 
@@ -61,9 +42,9 @@ function SignUp(props) {
     };
 
     return (
-        <Container component="main" maxWidth="xs">
+        <Container maxWidth="xs">
             <CssBaseline />
-            <div className={classes.paper}>
+            <Paper elevation={3} className={classes.paper}>
                 <Avatar className={classes.avatar}>
                     <LockOutlinedIcon />
                 </Avatar>
@@ -202,7 +183,7 @@ function SignUp(props) {
                         </Form>
                     }
                 </Formik>
-            </div>
+            </Paper>
             <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
                 <MuiAlert onClose={handleClose} severity={alertSeverity} elevation={6} variant="filled" >
                     {alertMessage}
